requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
astunparse==1.6.3
attrs==25.3.0
backports-datetime-fromisoformat==2.0.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
dnspython==2.8.0
email-validator==2.3.0
environs==14.3.0
exceptiongroup==1.3.0
fastapi==0.115.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastapi_cors==0.0.6
flatbuffers==25.9.23
fonttools==4.60.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
jax==0.4.34
jaxlib==0.4.34
Jinja2==3.1.6
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==4.0.1
matplotlib==3.10.6
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.3.2
namex==0.1.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pillow==10.4.0
protobuf==4.25.3
pycparser==2.23
pydantic==2.9.2
pydantic-extra-types==2.10.5
pydantic-settings==2.11.0
pydantic_core==2.23.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
scipy==1.15.3
sentencepiece==0.2.1
sentry-sdk==2.39.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.38.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
